Fix logger and agent tests
