+++ pkg/webhook/admission/pod/agent_injector_test.go
@@ -17,6 +17,7 @@ limitations under the License.
 package pod
 
 import (
+	"github.com/kubeflow/kfserving/pkg/apis/serving/v1beta1"
 	"github.com/kubeflow/kfserving/pkg/credentials"
 	"k8s.io/apimachinery/pkg/api/resource"
 	"knative.dev/pkg/kmp"
@@ -25,6 +26,13 @@ import (
 	"github.com/kubeflow/kfserving/pkg/constants"
 	v1 "k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+)
+
+const (
+	AgentDefaultCPURequest    = "100m"
+	AgentDefaultCPULimit      = "1"
+	AgentDefaultMemoryRequest = "200Mi"
+	AgentDefaultMemoryLimit   = "1Gi"
 )
 
 var (
@@ -36,14 +44,18 @@ var (
 		MemoryLimit:   AgentDefaultMemoryLimit,
 	}
 
+	loggerConfig = &LoggerConfig{
+		Image: "gcr.io/kfserving/agent:latest",
+	}
+
 	agentResourceRequirement = v1.ResourceRequirements{
 		Limits: map[v1.ResourceName]resource.Quantity{
-			v1.ResourceCPU:    resource.MustParse(BatcherDefaultCPULimit),
-			v1.ResourceMemory: resource.MustParse(BatcherDefaultMemoryLimit),
+			v1.ResourceCPU:    resource.MustParse(AgentDefaultCPULimit),
+			v1.ResourceMemory: resource.MustParse(AgentDefaultMemoryLimit),
 		},
 		Requests: map[v1.ResourceName]resource.Quantity{
-			v1.ResourceCPU:    resource.MustParse(BatcherDefaultCPURequest),
-			v1.ResourceMemory: resource.MustParse(BatcherDefaultMemoryRequest),
+			v1.ResourceCPU:    resource.MustParse(AgentDefaultCPURequest),
+			v1.ResourceMemory: resource.MustParse(AgentDefaultMemoryRequest),
 		},
 	}
 )
@@ -108,6 +120,7 @@ func TestAgentInjector(t *testing.T) {
 								},
 							},
 							Args: []string{"-config-dir", "/mnt/configs", "-model-dir", "/mnt/models"},
+							Env:  []v1.EnvVar{},
 						},
 					},
 					Volumes: []v1.Volume{
@@ -153,6 +166,99 @@ func TestAgentInjector(t *testing.T) {
 				},
 			},
 		},
+		"AddLogger": {
+			original: &v1.Pod{
+				ObjectMeta: metav1.ObjectMeta{
+					Name:      "deployment",
+					Namespace: "default",
+					Annotations: map[string]string{
+						constants.LoggerInternalAnnotationKey:        "true",
+						constants.LoggerSinkUrlInternalAnnotationKey: "http://httpbin.org/",
+						constants.LoggerModeInternalAnnotationKey:    string(v1beta1.LogAll),
+					},
+					Labels: map[string]string{
+						"serving.kubeflow.org/inferenceservice": "sklearn",
+						constants.KServiceModelLabel:            "sklearn",
+						constants.KServiceEndpointLabel:         "default",
+						constants.KServiceComponentLabel:        "predictor",
+					},
+				},
+				Spec: v1.PodSpec{
+					Containers: []v1.Container{{
+						Name: "sklearn",
+					}},
+				},
+			},
+			expected: &v1.Pod{
+				ObjectMeta: metav1.ObjectMeta{
+					Name: "deployment",
+					Annotations: map[string]string{
+						constants.LoggerInternalAnnotationKey:        "true",
+						constants.LoggerSinkUrlInternalAnnotationKey: "http://httpbin.org/",
+						constants.LoggerModeInternalAnnotationKey:    string(v1beta1.LogAll),
+					},
+				},
+				Spec: v1.PodSpec{
+					Containers: []v1.Container{{
+						Name: "sklearn",
+					},
+						{
+							Name:  constants.AgentContainerName,
+							Image: loggerConfig.Image,
+							Args: []string{
+								LoggerArgumentLogUrl,
+								"http://httpbin.org/",
+								LoggerArgumentSourceUri,
+								"deployment",
+								LoggerArgumentMode,
+								"all",
+								LoggerArgumentInferenceService,
+								"sklearn",
+								LoggerArgumentNamespace,
+								"default",
+								LoggerArgumentEndpoint,
+								"default",
+							},
+							Env:       []v1.EnvVar{},
+							Resources: agentResourceRequirement,
+							ReadinessProbe: &v1.Probe{
+								Handler: v1.Handler{
+									Exec: &v1.ExecAction{
+										Command: []string{
+											"/agent",
+											"-probe-period",
+											"0",
+										},
+									},
+								},
+							},
+						},
+					},
+				},
+			},
+		},
+		"DoNotAddLogger": {
+			original: &v1.Pod{
+				ObjectMeta: metav1.ObjectMeta{
+					Name: "deployment",
+				},
+				Spec: v1.PodSpec{
+					Containers: []v1.Container{{
+						Name: "sklearn",
+					}},
+				},
+			},
+			expected: &v1.Pod{
+				ObjectMeta: metav1.ObjectMeta{
+					Name: "deployment",
+				},
+				Spec: v1.PodSpec{
+					Containers: []v1.Container{{
+						Name: "sklearn",
+					}},
+				},
+			},
+		},
 	}
 
 	credentialBuilder := credentials.NewCredentialBulder(c, &v1.ConfigMap{
@@ -163,6 +269,7 @@ func TestAgentInjector(t *testing.T) {
 		injector := &AgentInjector{
 			credentialBuilder,
 			agentConfig,
+			loggerConfig,
 		}
 		injector.InjectAgent(scenario.original)
 		if diff, _ := kmp.SafeDiff(scenario.expected.Spec, scenario.original.Spec); diff != "" {
